// packages/solutions-components/src/utils/interfaces.ts
export enum ESelectionMode {
  ADD="ADD",
  REMOVE="REMOVE"
}

export enum ESelectionType {
  POINT="POINT",
  LINE="LINE",
  POLY="POLY",
  RECT="RECT"
}

export enum EWorkflowType {
  SEARCH="SEARCH",
  SELECT="SELECT",
  SKETCH="SKETCH",
  REFINE="REFINE"
}

export enum EExportType {
  CSV="CSV",
  PDF="PDF"
}

export enum EDrawMode {
  SKETCH="SKETCH",
  REFINE="REFINE"
}

/* eslint-enable no-unused-vars */

export type ValidSize = 6|10|14|20|30|60|80;

export type DistanceUnit = "feet"|"meters"|"miles"|"kilometers";
// ...
export interface IRefineOperation {
  ids: number[];
  mode: ESelectionMode;
  layerView: __esri.FeatureLayerView;
}
// ...
export interface ISelectionSet {
  id: number; // Date.Now() when the item is created...used to update a selection set
  searchResult: any;
  buffer: __esri.Geometry;
  distance: number;
  download: boolean;
  unit: DistanceUnit;
  label: string;
  selectedIds: number[];
  layerView: __esri.FeatureLayerView;
  geometries: __esri.Geometry[];
  graphics: __esri.Graphic[];
  selectLayers: __esri.FeatureLayerView[];
  skipGeomOIDs?: number[];
  workflowType: EWorkflowType;
  searchDistanceEnabled: boolean;
  useLayerFeaturesEnabled: boolean;
  refineInfos: IRefineInfo;
  redoStack?: IRefineOperation[];
  undoStack?: IRefineOperation[];
  sketchGraphic: __esri.Graphic; // See: https://github.com/Esri/solutions-components/issues/208
}

export interface IRefineInfo {
  [key: string]: IRefineIds;
}

export interface IRefineIds {
  addIds: number[];
  removeIds: number[];
  layerView: __esri.FeatureLayerView;
}